Recycle View y database propiamente utilizado TODO: sincronizar info
