+++ app/src/main/java/ado/edu/pucmm/rancherasystem/ConsultarCliente.java
@@ -19,3 +19,5 @@ public class ConsultarCliente extends AppCompatActivity {
         myToast.show();
     }
 }
+
+
