+++ app/src/main/java/ado/edu/pucmm/rancherasystem/ListOfBills.java
@@ -1,13 +1,18 @@
 package ado.edu.pucmm.rancherasystem;
 
+import android.arch.lifecycle.Observer;
+import android.arch.lifecycle.ViewModelProviders;
 import android.arch.persistence.db.SupportSQLiteDatabase;
 import android.arch.persistence.room.RoomDatabase;
 import android.os.AsyncTask;
 import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
+
+import java.util.List;
 
 import ado.edu.pucmm.rancherasystem.db.Bill;
 import ado.edu.pucmm.rancherasystem.db.BillDao;
@@ -15,7 +20,7 @@ import ado.edu.pucmm.rancherasystem.db.RancheraDB;
 
 public class ListOfBills extends AppCompatActivity {
 
-
+    private BillViewModel mBillViewModel;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -24,36 +29,22 @@ public class ListOfBills extends AppCompatActivity {
 
         RecyclerView recyclerView = findViewById(R.id.recyclerview);
         final BillListAdapter adapter = new BillListAdapter(this);
+
+        /*
+        mBillViewModel = ViewModelProviders.of(this).get(BillViewModel.class);
+
+        mBillViewModel.getmAllBills().observe(this, new Observer<List<Bill>>() {
+            @Override
+            public void onChanged(@Nullable final List<Bill> bills) {
+                // Update the cached copy of the words in the adapter.
+                adapter.setBills(bills);
+            }
+        }); */
+
         recyclerView.setAdapter(adapter);
         recyclerView.setLayoutManager(new LinearLayoutManager(this));
+
+
     }
 
-    private static RoomDatabase.Callback sRoomDatabaseCallback =
-            new RoomDatabase.Callback(){
-
-                @Override
-                public void onOpen (@NonNull SupportSQLiteDatabase db){
-                    super.onOpen(db);
-                    new PopulateDbAsync(INSTANCE).execute();
-                }
-            };
-
-    private static class PopulateDbAsync extends AsyncTask<Void, Void, Void> {
-
-        private final BillDao mDao;
-
-        PopulateDbAsync(RancheraDB db) {
-            mDao = db.billDao();
-        }
-
-        @Override
-        protected Void doInBackground(final Void... params) {
-            mDao.deleteAll();
-            Bill bill = new Bill(1, 500, "Primer test de esto", 1);
-            mDao.insert(bill);
-            bill = new Bill(2,5000,"segundo test", 1);
-            mDao.insert(bill);
-            return null;
-        }
-    }
 }
