+++ app/src/main/java/ado/edu/pucmm/rancherasystem/MainActivity.java
@@ -4,14 +4,16 @@ import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
+import android.widget.Button;
 
 public class MainActivity extends AppCompatActivity {
 
+    String id_client;
+
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
+    protected void onCreate(final Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-
 
     }
 
@@ -32,4 +34,11 @@ public class MainActivity extends AppCompatActivity {
         //login logic
         startActivity(helpIntent);
     }
+
+
+
+
+
+
+
 }
