+++ app/src/main/java/ado/edu/pucmm/rancherasystem/db/RancheraDB.java
@@ -5,6 +5,7 @@ import android.arch.persistence.room.Database;
 import android.arch.persistence.room.Room;
 import android.arch.persistence.room.RoomDatabase;
 import android.content.Context;
+import android.os.AsyncTask;
 import android.support.annotation.NonNull;
 
 @Database(entities = {Client.class, Bill.class, Product.class}, version = 1)
@@ -21,6 +22,7 @@ public abstract class RancheraDB extends RoomDatabase {
                 if (INSTANCE == null) {
                     INSTANCE = Room.databaseBuilder(context.getApplicationContext(),
                             RancheraDB.class, "ranchera_database")
+                            .addCallback(sRoomDatabaseCallback)
                             .build();
                 }
             }
@@ -32,13 +34,33 @@ public abstract class RancheraDB extends RoomDatabase {
         INSTANCE = null;
     }
 
+
     private static RoomDatabase.Callback sRoomDatabaseCallback =
             new RoomDatabase.Callback(){
 
                 @Override
                 public void onOpen (@NonNull SupportSQLiteDatabase db){
                     super.onOpen(db);
-                    //new PopulateDbAsync(INSTANCE).execute();
+                    new PopulateDbAsync(INSTANCE).execute();
                 }
             };
+
+    private static class PopulateDbAsync extends AsyncTask<Void, Void, Void> {
+
+        private final BillDao mDao;
+
+        PopulateDbAsync(RancheraDB db) {
+            mDao = db.billDao();
+        }
+
+        @Override
+        protected Void doInBackground(final Void... params) {
+            mDao.deleteAll();
+            Bill bill = new Bill(1, 500, "Primer test de esto", 1);
+            mDao.insert(bill);
+            bill = new Bill(2,5000,"segundo test", 1);
+            mDao.insert(bill);
+            return null;
+        }
+    }
 }
