+++ app/src/main/res/layout/activity_dash_board.xml
@@ -6,4 +6,5 @@
     android:layout_height="match_parent"
     tools:context=".DashBoardActivity">
 
+
 </android.support.constraint.ConstraintLayout>
