+++ app/src/main/res/layout/fragment_ordenes_page.xml
@@ -10,5 +10,6 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:text="Ordenes Page" />
+    
 
 </FrameLayout>
